Rename editor module to "programs"
This module will be extended to provide convenient access to programs
other than the editor.
